# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: RentFlow API
  description: |
    Multi-tenant Rent Management SaaS API.

    **Rules enforced at every endpoint:**
    - `public_id` is used in all URLs and responses — internal `id` is never exposed
    - All resources are scoped to the authenticated owner via JWT
    - Soft-deleted resources (`is_active=false`) are treated as not found (404)
    - All monetary values are numbers (`NUMERIC(10,2)` precision)
  version: "1.0.0"
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # --------------------------------------------------------------
  # Reusable path parameters
  # --------------------------------------------------------------
  parameters:
    page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    page_size:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    building_public_id:
      name: building_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    tenant_public_id:
      name: tenant_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Authentication required"
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Access denied"
    NotFound:
      description: Resource not found or soft-deleted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Not found"
    BadRequest:
      description: Business logic error (e.g. duplicate unit, conflicting dates, default category delete)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'

  # --------------------------------------------------------------
  # Schemas
  # --------------------------------------------------------------
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
          example: null
        message:
          type: string
    ValidationErrorEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: false
        data:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        message:
          type: string
          example: "Validation failed"
    Pagination:
      type: object
      required: [total, page, page_size]
      properties:
        total:
          type: integer
          description: Total number of matching records
        page:
          type: integer
        page_size:
          type: integer

    # ── Auth ───────────────────────────────────────────────────
...
    BuildingCreate:
      type: object
      required: [name, address, total_floors]
      properties:
        name:
          type: string
          maxLength: 255
        address:
          type: string
        total_floors:
          type: integer
          minimum: 1
    BuildingUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        address:
          type: string
        total_floors:
          type: integer
          minimum: 1
    BuildingData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        total_floors:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Apartments ─────────────────────────────────────────────
...
    DueGenerateRequest:
      type: object
      required: [month, year]
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
          minimum: 2000
          maximum: 2100
        due_date:
          type: string
          format: date
          nullable: true
          description: "Custom due date. Defaults to the 1st of the month if omitted."
...
    MonthlyDueData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        tenant_public_id:
          type: string
          format: uuid
        agreement_public_id:
          type: string
          format: uuid
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        rent_amount:
          type: number
        total_due:
          type: number
        amount_paid:
          type: number
        remaining_balance:
          type: number
        status:
          type: string
          enum: [unpaid, partial, paid]
        is_auto_generated:
          type: boolean
        due_date:
          type: string
          format: date
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Payments ───────────────────────────────────────────────
...
paths:
...
  /buildings:
    get:
      tags: [Buildings]
      summary: List all buildings for the authenticated owner
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
      responses:
        "200":
          description: Buildings list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/BuildingData'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Buildings]
      summary: Create a new building
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildingCreate'
      responses:
        "201":
          description: Building created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/BuildingData'
                  message: { type: string, example: "Building created" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  /buildings/{building_public_id}:
    parameters:
      - $ref: '#/components/parameters/building_public_id'
    get:
      tags: [Buildings]
      summary: Get a single building
      responses:
        "200":
          description: Building details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/BuildingData'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Buildings]
      summary: Update building details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildingUpdate'
      responses:
        "200":
          description: Building updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/BuildingData'
                  message: { type: string, example: "Building updated" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Buildings]
      summary: Soft delete a building
      description: "Returns 400 if the building has active apartments."
      responses:
        "200":
          description: Building deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { nullable: true, example: null }
                  message: { type: string, example: "Building deactivated" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

  # ==============================================================
  # APARTMENTS
  # ==============================================================
...
  /tenants/{tenant_public_id}/dues:
    parameters:
      - $ref: '#/components/parameters/tenant_public_id'
    get:
      tags: [Monthly Dues]
      summary: List all monthly dues for a tenant
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
        - name: status
          in: query
          schema:
            type: string
            enum: [unpaid, partial, paid]
        - name: year
          in: query
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
      responses:
        "200":
          description: Monthly dues list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MonthlyDueData'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /tenants/{tenant_public_id}/dues/generate:
    parameters:
      - $ref: '#/components/parameters/tenant_public_id'
    post:
      tags: [Monthly Dues]
      summary: Auto-generate a monthly due from the active agreement
      description: |
        Reads the active agreement's rent_amount and creates a MonthlyDue.
        Returns 400 if a due for (tenant, month, year) already exists or if there
        is no active agreement.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DueGenerateRequest'
      responses:
        "201":
          description: Monthly due generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/MonthlyDueData'
                  message: { type: string, example: "Due generated" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
...